requirements.txt
Django>=6.0,<7.0
Pillow>=10.0
dj-database-url>=2.2
gunicorn>=22.0
psycopg[binary]>=3.2
whitenoise>=6.7
